+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -80,6 +80,10 @@ New checks
   Finds non-const global variables as described in check I.2 of C++ Core
   Guidelines.
 
+- New :doc:`bsl-literals-uppercase-suffix
+  <clang-tidy/checks/bsl-literals-uppercase-suffix>` check.
+
+  Checks that each integer literal has an uppercase integer suffix.
 - New :doc:`bugprone-misplaced-pointer-arithmetic-in-alloc
   <clang-tidy/checks/bugprone-misplaced-pointer-arithmetic-in-alloc>` check.